wire up Firestore database and add a toggle for syncing job application. this toggle is disabled for offline user.

+++ app/src/main/java/com/jobapplicationapp/jobby/data/SyncJobApplicationRepository.kt
@@ -2,12 +2,15 @@ package com.jobapplicationapp.jobby.data
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.firstOrNull
 import kotlinx.coroutines.launch
 
 /**
  * A repository for coordinates local (Room) and remote (Firestore) data sources.
  * It follows the Offline-First principle: UI to always reads from local data.
  * Changes are saved locally first, then synced to remote if enabled.
+ * Note that both localRepository and remoteRepostiroy variable is set to JobApplicationRepository
+ * for interface
  */
 class SyncJobApplicationRepository(
     private val localRepository: JobApplicationRepository,
@@ -15,9 +18,25 @@ class SyncJobApplicationRepository(
     private val externalScope: CoroutineScope
 ) : JobApplicationRepository {
 
-    var isSyncEnabled: Boolean = false // flag based on user setting
+    var isSyncEnabled: Boolean = false
+    private var isSyncing: Boolean = false
 
     override fun getAllJobApplications(userId: String): Flow<List<JobApplication>> {
+        if (isSyncEnabled && userId != OFFLINE_USER_ID) {
+            externalScope.launch {
+                if (!isSyncing) {
+                    isSyncing = true
+                    val getJobsFromRemote = remoteRepository.getAllJobApplications(userId)
+                    getJobsFromRemote.collect { remoteJobs ->
+                        remoteJobs.forEach { remoteJob ->
+                            remoteRepository.updateJobApplication(remoteJob.copy(isSynced = true))
+                            localRepository.updateJobApplication(remoteJob)
+                        }
+                    }
+                    isSyncing = false
+                }
+            }
+        }
         return localRepository.getAllJobApplications(userId)
     }
 
@@ -28,7 +47,7 @@ class SyncJobApplicationRepository(
         )
         localRepository.addJobApplication(jobWithMetadata)
         
-        if (isSyncEnabled) {
+        if (isSyncEnabled && jobApplication.userId != OFFLINE_USER_ID) {
             trySync(jobWithMetadata)
         }
     }
@@ -39,7 +58,7 @@ class SyncJobApplicationRepository(
 
     override suspend fun deleteJobApplication(jobApplication: JobApplication) {
         localRepository.deleteJobApplication(jobApplication)
-        if (isSyncEnabled) {
+        if (isSyncEnabled && jobApplication.userId != OFFLINE_USER_ID) {
             externalScope.launch {
                 try {
                     remoteRepository.deleteJobApplication(jobApplication)
